Core improvements, renderer and resource updates, bug fixes, and minor adjustments
• Added new resource files to support rendering and gameplay.
• Added Depth struct for consistent depth layer handling in rendering.
• Enhanced the Sprite and Texture classes with better encapsulation and performance.
• Fixed bug in the camera's view matrix calculation and corrected lighting in Lit.glsl.
• Improved the Renderer class with better structure and functionality.
• Introduced Allocators module with abstracted smart pointer wrappers.
• Refined engine primitive types for improved type safety and consistency.
• Replaced all raw C++ primitive types with Lion Engine's custom primitive types across the codebase for consistency.

+++ Engine/Source/Render/Camera.h
@@ -10,7 +10,7 @@ namespace Lion
 		virtual const LION_API glm::mat4& GetViewMatrix() const { return mView; }
 		virtual const LION_API glm::mat4& GetProjectionMatrix() const { return mProjection; }
 
-		virtual LION_API void OnResize(int width, int height) = 0;
+		virtual LION_API void OnResize(float32 width, float32 height) = 0;
 		virtual LION_API void OnUsage() = 0;
 
 	protected:
@@ -21,9 +21,9 @@ namespace Lion
 		glm::vec3 mRotation;
 		const glm::vec3 mUp = glm::vec3{ 0.0f, 1.0f, 0.0f };
 
-		float mAspectRatio;
-		float mNearPlane;
-		float mFarPlane;
-		float mFieldOfView;
+		float32 mAspectRatio;
+		float32 mNearPlane;
+		float32 mFarPlane;
+		float32 mFieldOfView;
 	};
 }
